Add experimental support for compiling ksh93 with PGO

+++ PKGBUILD
@@ -4,7 +4,7 @@
 
 _pkgname='ksh93'
 pkgname="${_pkgname}-git"
-pkgver=r1807.8ee1efd9
+pkgver=r1824.57f51b62
 pkgrel=1
 pkgdesc="KornShell 93u+m, fork based on ksh 93u+"
 arch=('x86_64' 'i686' 'pentium4' 'powerpc64le' 'powerpc64' 'powerpc' 'riscv64' 'arm' 'armv6h' 'armv7h' 'aarch64')
@@ -12,6 +12,7 @@ url='https://github.com/ksh93/ksh/'
 license=('EPL')
 depends=('glibc')
 makedepends=('git')
+optdepends=('tcc: for faster PGO compilation (PGO is disabled by default)')
 conflicts=('ksh' 'ksh93')
 provides=('ksh' 'ksh93')
 install='ksh93.install'
@@ -27,9 +28,8 @@ pkgver() {
 
 prepare() {
 	cd "${srcdir}/${_pkgname}"
-	# Build with SHOPT_ALL_LIBCMD (aka enable all ksh builtins) if
-	# ${_all_libcmd} is set to '1' or 'yes'.
-	if [[ ${_all_libcmd} == 1 || ${_all_libcmd} == yes ]]; then
+	# Build with SHOPT_ALL_LIBCMD (aka enable all ksh builtins) if ${_all_libcmd} is enabled.
+	if [[ ${_all_libcmd} == 1 || ${_all_libcmd} == yes || ${_all_libcmd} == true ]]; then
 		sed -i 's/ALL_LIBCMD=0/ALL_LIBCMD=1/g' src/cmd/ksh93/SHOPT.sh
 	fi
 	# 'bin/package install' requires that the file(1) command be functional.
@@ -42,9 +42,54 @@ prepare() {
 
 build() {
 	cd "${srcdir}/${_pkgname}"
-	rm -rf ./arch               # Get rid of any leftover build files (in case makepkg.conf changed)
-	export CCFLAGS="${CFLAGS}"  # bin/package uses CCFLAGS rather than CFLAGS
-	./bin/package make          # Build ksh
+	# Get rid of any leftover build files (in case makepkg.conf changed).
+	rm -rf ./arch
+	# Obtain the number of CPU cores.
+	local -i cores=$(bin/package host cpu)
+	if ! [[ ${_ksh_pgo} == 1 || ${_ksh_pgo} == yes || ${_ksh_pgo} == true ]]; then
+		export CCFLAGS=${CFLAGS}  # bin/package uses CCFLAGS rather than CFLAGS.
+		./bin/package make        # Build ksh (no -j flag because that's still experimental).
+	else
+		# Optional and experimental: Compile ksh with profile guided optimization (making
+		# use of the regression tests) if ${_ksh_pgo} is enabled.
+		local save_ccflags=${CFLAGS}
+		local save_ldflags=${LDFLAGS}
+		# The first build is only done to obtain a PGO-less copy of the pty executable,
+		# such that an interactive ksh can be profiled without contaminating the
+		# profiling data with useless pty heuristics.
+		# (Additionally, tcc is used when available to increase compilation speed).
+		if type tcc >/dev/null 2>&1; then
+			unset LDFLAGS  # CCFLAGS isn't set yet, so don't bother to unset it
+			./bin/package make -j${cores} CC=tcc
+		else
+			./bin/package make -j${cores}
+		fi
+		mv arch/*/bin/pty ./saved-pty
+		rm -rf ./arch
+		# Create a temporary directory to use for PGO
+		local tmpdir="${PWD}/pgotmp-$SRANDOM"
+		mkdir "$tmpdir"
+		# Second build with profiling flags set (-fno-unroll-loops increases overall
+		# performance slightly according to my results from shbench).
+		local generation_flags="-fprofile-dir=\"${tmpdir}\" -fprofile-generate=\"${tmpdir}\" -fno-unroll-loops"
+		local use_flags="-fprofile-dir=\"${tmpdir}\" -fprofile-use=\"${tmpdir}\" -fprofile-correction -fno-unroll-loops -Wno-error=coverage-mismatch"
+		export CCFLAGS="${save_ccflags} ${generation_flags}"
+		export LDFLAGS="${save_ldflags} ${generation_flags}"
+		bin/package make -j1
+		# Discard irrelevant profiling data from mamake
+		rm -r "${tmpdir}" && mkdir "${tmpdir}"
+		# Run the regression tests to profile ksh
+		cp ./saved-pty arch/*/bin/pty
+		rm arch/*/dyn/bin/pty
+		./arch/*/bin/ksh bin/shtests --posix || true
+		# Final stage of PGO
+		export CCFLAGS="${save_ccflags} ${use_flags}"
+		export LDFLAGS="${save_ldflags} ${use_flags}"
+		rm -rf ./arch
+		./bin/package make -j${cores}
+		# Cleanup PGO artefacts
+		rm -r "${tmpdir}" saved-pty
+	fi
 }
 
 package() {
@@ -65,12 +110,14 @@ package() {
 	# Install license files and additional documentation
 	install -Dm0644 LICENSE.md "${pkgdir}/usr/share/licenses/ksh/LICENSE.md"
 	install -Dm0644 COPYRIGHT "${pkgdir}/usr/share/licenses/ksh/COPYRIGHT"
+	local _doc
 	for _doc in 'builtins.mm' 'COMPATIBILITY' 'DESIGN' 'OBSOLETE' 'PROMO.mm' 'README' 'README-AUDIT.md' \
 			'RELEASE' 'RELEASE88' 'RELEASE93' 'sh.memo' 'TYPES'; do
 		install -Dm0644 "src/cmd/${_pkgname}/${_doc}" "${pkgdir}/usr/share/doc/ksh/${_doc}"
 	done
 
 	# Create relevant relative symlinks for alternate ksh names
+	local _kshname
 	for _kshname in 'ksh93' 'rksh' 'rksh93'; do
 		ln -sr "${pkgdir}/usr/bin/ksh" "${pkgdir}/usr/bin/${_kshname}"
 		ln -sr "${pkgdir}/usr/share/man/man1/ksh.1" "${pkgdir}/usr/share/man/man1/${_kshname}.1"
